move certain components and libs under packages (not publishable yet)
