+++ components/PublicationTabs.tsx
@@ -6,13 +6,13 @@ import BiospecimenTable from './BiospecimenTable';
 import CaseTable from './CaseTable';
 import FileTable from './FileTable';
 import { DataSchemaData } from '../lib/dataSchemaHelpers';
-import { groupEntitiesByAttrNameAndValue } from '../lib/filterHelpers';
 import { Atlas, Entity, setTab } from '../lib/helpers';
 import styles from './PublicationTabs.module.scss';
 import { Badge } from 'react-bootstrap';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faExternalLinkAlt } from '@fortawesome/free-solid-svg-icons';
 import { FileAttributeMap, GenericAttributeNames } from '../lib/types';
+import { groupEntitiesByAttrNameAndValue } from '../packages/data-portal-filter/src/libs/helpers';
 
 interface IPublicationTabsProps {
     router: NextRouter;
