+++ packages/data-portal-utils/src/libs/clamp.ts
@@ -1,0 +1,3 @@
+export function clamp(x: number, lower: number, upper: number) {
+    return Math.max(lower, Math.min(x, upper));
+}
