+++ packages/data-portal-utils/src/libs/types.ts
@@ -1,0 +1,6 @@
+export interface IAttributeInfo<T> {
+    path?: string;
+    getValues?: (e: T) => string[];
+    displayName: string;
+    caseFilter?: boolean;
+}
