+++ pages/tools.tsx
@@ -3,12 +3,10 @@ import PreReleaseBanner from '../components/PreReleaseBanner';
 import PageWrapper from '../components/PageWrapper';
 import { getToolData, Tools } from '../lib/tools';
 import ToolTable from '../components/ToolTable';
-import Filter from '../components/filter/Filter';
 import React from 'react';
 import ToolFilterControls from '../components/filter/ToolFilterControls';
 import { observer } from 'mobx-react';
 import { NextRouter, withRouter } from 'next/router';
-import { ExploreActionMeta, ExploreSelectedFilter } from '../lib/types';
 import {
     parseSelectedFiltersFromUrl,
     updateSelectedFiltersInURL,
@@ -16,24 +14,30 @@ import {
 import { ExploreURLQuery } from './explore';
 import {
     filtertools,
-    getNewFilters,
-    getSelectedFiltersByAttrName,
     getToolFilterDisplayName,
     groupToolsByAttrNameAndValue,
 } from '../lib/filterHelpers';
 import { action } from 'mobx';
+
+import {
+    FilterActionMeta,
+    SelectedFilter,
+} from '../packages/data-portal-filter/src/libs/types';
+import {
+    getNewFilters,
+    getSelectedFiltersByAttrName,
+} from '../packages/data-portal-filter/src/libs/helpers';
+import Filter from '../packages/data-portal-filter/src/components/Filter';
 
 const ToolPage = observer((props: { router: NextRouter; tools: Tools }) => {
     const selectedFilters =
         parseSelectedFiltersFromUrl(
             (props.router.query as ExploreURLQuery).selectedFilters // use casting as ExploreURLQuery to use typescript to ensure URL correctness
         ) || [];
-    const setFilter = action(
-        (actionMeta: ExploreActionMeta<ExploreSelectedFilter>) => {
-            const newFilters = getNewFilters(selectedFilters, actionMeta);
-            updateSelectedFiltersInURL(newFilters, props.router);
-        }
-    );
+    const setFilter = action((actionMeta: FilterActionMeta<SelectedFilter>) => {
+        const newFilters = getNewFilters(selectedFilters, actionMeta);
+        updateSelectedFiltersInURL(newFilters, props.router);
+    });
     const selectedFiltersByAttrName = getSelectedFiltersByAttrName(
         selectedFilters
     );
